Akaro second commit - More pages have been added; - Added new resources and colors; - Current card, sign in and sign up pages were made; - Filtering by product categories has been added; - Authorization and registration via Firebase were implemented also adding data to the room db.

+++ app/src/main/java/com/example/beautifulprincess/models/ProductRowModel.kt
@@ -2,6 +2,6 @@ package com.example.beautifulprincess.models
 
 class ProductRowModel (
     val title: String,
-    val price: Double,
+    val price: Float,
     val imageId: Int
 )
